business tests

+++ Entities/Concrete/EmployerRating.cs
@@ -7,7 +7,7 @@ namespace Entities.Concrete
     {
 		public int Id { get; set; }
 		public int EmployerId { get; set; }
-		public short Rating { get; set; }
+		public byte Rating { get; set; }
 		public string? Comment { get; set; }
 		public DateTime Date { get; set; }
 		public int UserId { get; set; }
